// src/lib/api/generated/models/packAttributeInputDtoMatchType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type PackAttributeInputDtoMatchType = typeof PackAttributeInputDtoMatchType[keyof typeof PackAttributeInputDtoMatchType];


export const PackAttributeInputDtoMatchType = {
  COMPATIBLE: 'COMPATIBLE',
  NOT_COMPATIBLE: 'NOT_COMPATIBLE',
  BOOST: 'BOOST',
} as const;